+++ database/migrations/2024_04_03_031335_create_account_transaction_detail_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('accountNo')->required();
             $table->foreign('accountNo')->references('accountNo')->on('account')->onDelete('restrict');
             $table->string('kind', 1)->required();
-            $table->bigInteger('total');
+            $table->decimal('total', 10, 2);
             $table->timestamps();
         });
     }
